# Improve prime sieve performance