--- a/base/exports.jl
+++ b/base/exports.jl
@@ -415,6 +415,7 @@ export
     prevpow2,
     prevprod,
     primes,
+    primesmask,
     rad2deg,
     rationalize,
     real,