This PR changes the algorithm used in `primesmask` from a Sieve of Atkin to a Sieve of Eratosthenes, based on the ~~idea~~ redesign proposed by @VicDrastik in #11594.

The main ideas used are

* ~~Most integers are composite, so it is better to set all flags to false initially, then invert for possible primes~~
* ~~Except for 2, 3 all primes are of the from 6k ± 1~~
* Implement a 30 wheel sieve. This sieve stores a flag for each number of the form 30k + p where p is one of `[7, 11, 13, 17, 19, 23, 29, 31]` (all primes greater than 5 are of this form).
* Cross-off multiples of p of the above form, starting from p².
* Use the wheel sieve `_primesmask` (not exported) to construct `primesmask` and to find primes.

Additionally, the function `primesmask` is now exported and both this function and `primes` are extended in order to generate all the primes between two integers.

As a follow-up, it would be worth trying to write a general segmented version (perhaps a mix between this and the ideas of @ironman353 in #11594 or this http://primesieve.org/segmented_sieve.html).

I apologize to @Ismael-VC who was also working on this in #11927 (his approach was to fix the current implementation of the Sieve of Atkin, but this should be even faster as suggested in #11594).

Finally, this PR addresses the initial issue on #11594 but that turned out to be a discussion on how to improve the whole support for primes. Maybe that one should be separated into smaller issues and grouped into a master issue.

CC @danaj @aiorla @StefanKarpinski