@pabloferz 2015-07-06
This still needs some tests, news and documentation.