@Ismael-VC 2015-07-06
This is great! @pabloferz no apologies needed, I'm still learning and this is what I wanted to see in order to learn more, thanks!