@Mike43110 2015-07-07
@ironman353 The Julia `Array` Type is the same as vector and ArrayList.