@pabloferz 2015-07-11
@ironman353 That optimized sieve of Atkin is really nice and is faster than what I had. Did you came up with it or did you get the algorithm from somewhere else? If it's yours would you mind if I rewrite the PR based on your code? If it's not, do you know if we can released it under the MIT license?