@pabloferz 2015-07-27
How can I find what conflicts this PR entails? Would a rebase fix that?

@StefanKarpinski