@KristofferC 2015-07-27
Rebase on master should fix yes.