@pabloferz 2015-07-29
I have a segmented version of the function `primes`, but for it to perform the best it would have to be cache-size dependent. But, I do not know how to get the cache size from within Julia or whether or not such kind of implementation is desirable. Besides that can be done in another PR.

It the proposed changes seem good, I can add the corresponding documentation so this can be merged.