@StefanKarpinski 2015-07-29
It would be quite nice to provide an API for finding out things like cache sizes.