@KristofferC 2015-07-29
Amen to that. I keep getting annoyed having to mentally parse the units. Before you just compared two numbers.