@mbauman 2015-07-29
There's still some known performance snags for BitArrays involving GC frames, pointer lookup hoists, and user-defined `@inbounds` (#11403, #11430, #7799).  I'm convinced we can eventually get them more competitive in more situations.

In this case, since BitArrays are within 15%, I'd try rewriting your `@inbounds` indexing with `unsafe_getindex` and `unsafe_setindex!`.  That might be enough to push BitArrays over the top.  That said, a 10x gain is *very* impressive!  Don't let me hold you up!