@pabloferz 2015-07-29
With `unsafe_setindex!` (the implementation does not uses `getindex`) the timings are these

```
julia> gc(); @time bitprimesmask(100_000_000);
 389.984 milliseconds (11 allocations: 38249 KB, 0.06% gc time)

julia> gc(); @time bitprimesmask(1_000_000_000);
   4.412 seconds      (11 allocations: 374 MB, 0.02% gc time)
```

This seems like a good compromise between speed and space. If there are plans on improve `BitArray`s performance, then there is another reason to use them for the `primesmask` function.