@tkelman 2015-07-31
Best to squash out any intermediate commits that would have failed tests.