@pabloferz 2015-08-01
I believe this PR is ready, so I leave it to you for reviewing it.

Just as a remark, the use of  a `BitArray` in the `primesmask` function does not affect the performance of  `primes`.