@pabloferz 2015-08-02
Also, this would close #11594, but a lot of the discussion there would have to be moved into at least another issue.