@pabloferz 2015-08-03
If I got #11943 right, this must be now using the new doc system and, hopefully, ready.