@tkelman 2015-08-03
I think so. You can run `./julia doc/genstdlib.jl` and see if the rst docs come out the way you'd like.