@pabloferz 2015-08-04
This PR, in my opinion, is ready to be merged, unless we need to wait for #12435 to land first.