@tkelman 2015-08-07
Is `primesmask` ending up in the generated RST docs? If not, I think you need to add a signature for it so `doc/genstdlib.jl` can fill in the docstring.