@pabloferz 2015-08-07
@tkelman No, it doesn't show un the RST docs. Where do I put the signature?