@tkelman 2015-08-07
Where in the manual do you want this function to be documented?