@pabloferz 2015-08-07
Justo under the current `primes` function under the Numbers section of the stdlib.