@pabloferz 2015-08-08
Ok, I tested it and it now generates the proper RST docs.