@tkelman 2015-08-08
@StefanKarpinski I'm going to assume you're cool with merging this?